[Feature] RunningTable 생성 시 HashTag 데이터 추가 완료
Former-commit-id: d38b0994b4dd03eb02976443fccc1827662cd268

+++ server/src/main/java/com/runstory/api/request/RunningCrewReqDto.java
@@ -43,5 +43,5 @@ public class RunningCrewReqDto {
     private boolean hasDog;
 
     // SelectedHashTag에 들어가기 위한 dto
-    private List<Integer> hastag;
+    private List<Long> hastag;
 }
